Changing script per Allie's feedback
Feedback:
I don't usually use dontdestroyonload but it's not wrong, just unneeded  awake can just be instance=this
instead of delta time, use an IndependentTimer (it's a class in the project). It has Start() and Finished() functions and you can get a 0-1 value to use for fades using Parameterized. Using just a regular Timer would be bad because of fish hits slowing down time for hitstop
What happens when you repeatedly press p/SimulateCombat()? Bad things lmfao. Instead of it being a coroutine (which have this problem of multiple happening at once), do everything in Update() manually so there's only one thing happening at a time

Changes awake to be instance = this
moved the simulate combat to the update function, and pressing P only has a little weirdness now if you spam it

I think i implemented the indepent timer correctly. Ill be honest, kinda melted my brain for the first 10 minutes